utils: move edit utils to tiup/utils
