+++ pkg/utils/diff_test.go
@@ -8,11 +8,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package edit
+package utils
 
 import (
-	"testing"
-
 	. "github.com/pingcap/check"
 	"gopkg.in/yaml.v2"
 )
@@ -21,10 +19,6 @@ type diffSuite struct {
 }
 
 var _ = Suite(&diffSuite{})
-
-func TestDiff(t *testing.T) {
-	TestingT(t)
-}
 
 type sampleDataMeta struct {
 	IntSlice     []int                    `yaml:"ints,omitempty"`
